# apps/backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de aplicacoes e deploys automaticos",
        "title": "PaaS Deploy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@paasdeploy.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/paas-deploy/v1",
    "paths": {
        "/apps": {
            "get": {
                "description": "Retorna lista de apps cadastrados no sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Lista todas as aplicacoes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.App"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            },
            "post": {
                "description": "Cadastra um novo app para deploy automatico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Cria uma nova aplicacao",
                "parameters": [
                    {
                        "description": "Dados do app",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.CreateAppInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/docs.App"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/apps/{id}": {
            "get": {
                "description": "Retorna os detalhes de um app especifico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Busca uma aplicacao por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do app",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.App"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deleta um app e remove o webhook associado",
                "tags": [
                    "apps"
                ],
                "summary": "Remove uma aplicacao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do app",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/apps/{id}/deployments": {
            "get": {
                "description": "Retorna historico de deploys de um app",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Lista deploys de uma aplicacao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do app",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.Deployment"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/apps/{id}/redeploy": {
            "post": {
                "description": "Inicia um deploy manual do app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Dispara um novo deploy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do app",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Commit SHA opcional",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.RedeployInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/docs.Deployment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/apps/{id}/rollback": {
            "post": {
                "description": "Reverte para o ultimo deploy bem-sucedido",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Faz rollback do deploy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do app",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/docs.Deployment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/apps/{id}/webhook": {
            "post": {
                "description": "Cria webhook automatico no repositorio GitHub",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Configura webhook do GitHub",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do app",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SetupResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deleta webhook do repositorio GitHub",
                "tags": [
                    "apps"
                ],
                "summary": "Remove webhook do GitHub",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do app",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/apps/{id}/webhook/status": {
            "get": {
                "description": "Retorna o status do webhook configurado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Verifica status do webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do app",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.WebhookStatus"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.App": {
            "description": "Aplicacao cadastrada para deploy automatico",
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string",
                    "example": "main"
                },
                "config": {
                    "type": "object"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "lastDeployedAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "my-app"
                },
                "repositoryUrl": {
                    "type": "string",
                    "example": "https://github.com/owner/repo.git"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive",
                        "deleted"
                    ],
                    "example": "active"
                },
                "updatedAt": {
                    "type": "string"
                },
                "webhookId": {
                    "type": "integer",
                    "example": 123456789
                },
                "workdir": {
                    "type": "string",
                    "example": "."
                }
            }
        },
        "docs.CreateAppInput": {
            "description": "Dados necessarios para cadastrar um novo app",
            "type": "object",
            "required": [
                "name",
                "repositoryUrl"
            ],
            "properties": {
                "branch": {
                    "type": "string",
                    "example": "main"
                },
                "config": {
                    "type": "object"
                },
                "name": {
                    "type": "string",
                    "example": "my-app"
                },
                "repositoryUrl": {
                    "type": "string",
                    "example": "https://github.com/owner/repo.git"
                },
                "workdir": {
                    "type": "string",
                    "example": "."
                }
            }
        },
        "docs.Deployment": {
            "description": "Registro de um deploy realizado",
            "type": "object",
            "properties": {
                "appId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "commitMessage": {
                    "type": "string",
                    "example": "feat: add new feature"
                },
                "commitSha": {
                    "type": "string",
                    "example": "abc123def456"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentImageTag": {
                    "type": "string"
                },
                "errorMessage": {
                    "type": "string"
                },
                "finishedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "logs": {
                    "type": "string"
                },
                "previousImageTag": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed",
                        "cancelled"
                    ],
                    "example": "success"
                }
            }
        },
        "docs.ErrorInfo": {
            "description": "Detalhes do erro",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "details": {},
                "message": {
                    "type": "string",
                    "example": "resource not found"
                }
            }
        },
        "docs.SetupResult": {
            "description": "Resultado da configuracao do webhook",
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "provider": {
                    "type": "string",
                    "example": "github"
                },
                "webhookId": {
                    "type": "integer",
                    "example": 123456789
                }
            }
        },
        "docs.WebhookStatus": {
            "description": "Status atual do webhook configurado",
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "error": {
                    "type": "string"
                },
                "exists": {
                    "type": "boolean",
                    "example": true
                },
                "lastPingAt": {
                    "type": "string"
                }
            }
        },
...
        "handler.RedeployInput": {
            "type": "object",
            "properties": {
                "commitSha": {
                    "type": "string",
                    "example": "abc123def"
                }
            }
        }
    },
...
}